+++ src/vroom_write.cc
@@ -76,9 +76,6 @@ size_t get_buffer_size(
 }
 
 bool needs_quote(const char* str, const char delim, const char* na_str) {
-  if (strncmp(str, na_str, 2) == 0) {
-    return true;
-  }
 
   for (const char* cur = str; *cur != '\0'; ++cur) {
     if (*cur == '\n' || *cur == '\r' || *cur == '"' || *cur == delim) {